Fixed GScore and HScore, RG Approach, slightly optimized, auto is smooth
Gscore and HScore have proper headings, RG Approach is a lot slower (for now) to avoid problems with elevator and deceleration. I also added a few extra rotation targets to make the turning smoother, and I fixed the paths so they are still out of the way for the most part and they are smoother and quicker. Lastly, I fixed GReferee to prevent us blocking other bots from hitting us near the station